split handler package
