+++ router.go
@@ -2,11 +2,11 @@ package main
 
 import (
 	"io"
-	"net/http"
 	"text/template"
 
 	"github.com/labstack/echo"
 	"github.com/labstack/echo/middleware"
+	"github.com/tomato3713/cw_for_web/handler"
 )
 
 type Template struct {
@@ -22,10 +22,16 @@ func newRouter() *echo.Echo {
 	e.Static("/assets", "static/assets")
 	e.File("/manifest.json", "static/manifest.json")
 	e.File("/favicon.ico", "static/assets/img/icon/favicon.ico")
-	e.GET("/", indexHandler)
-	e.GET("/runner", cwWebRunnerHandler)
-	e.GET("/tutor", cwWebTutorHandler)
-	e.GET("/history", historyHandler)
+	e.GET("/", handler.Index)
+	e.GET("/runner", handler.Runner)
+	e.GET("/tutor", handler.Tutor)
+	e.GET("/history", handler.History)
+
+	// authorization
+	e.GET("/signup", handler.Signup)
+	e.POST("/signup", handler.SignupProcess)
+	e.GET("/login", handler.Login)
+	e.POST("/login", handler.LoginProcess)
 
 	t := &Template{
 		templates: template.Must(template.ParseGlob("static/views/*.html")),
@@ -37,44 +43,4 @@ func newRouter() *echo.Echo {
 
 func (t *Template) Render(w io.Writer, name string, data interface{}, c echo.Context) error {
 	return t.templates.ExecuteTemplate(w, name, data)
-}
-
-func indexHandler(c echo.Context) error {
-	data := struct {
-		Info
-	}{
-		Info: info,
-	}
-
-	return c.Render(http.StatusOK, "index", data)
-}
-
-func cwWebRunnerHandler(c echo.Context) error {
-	data := struct {
-		Info
-	}{
-		Info: info,
-	}
-
-	return c.Render(http.StatusOK, "runner", data)
-}
-
-func cwWebTutorHandler(c echo.Context) error {
-	data := struct {
-		Info
-	}{
-		Info: info,
-	}
-
-	return c.Render(http.StatusOK, "tutor", data)
-}
-
-func historyHandler(c echo.Context) error {
-	data := struct {
-		Info
-	}{
-		Info: info,
-	}
-
-	return c.Render(http.StatusOK, "history", data)
 }
