+++ server.go
@@ -1,14 +1,5 @@
 package main
 
-// Info is common information on the service
-type Info struct {
-	Title string
-}
-
-var info = Info{
-	"CW Web Runner",
-}
-
 func main() {
 	router := newRouter()
